okay i found the little bug detail missing
